refactor : 공통되는 entity 속성 extends와 @MappedSuperclass 사용하여 공통화
