+++ src/main/java/com/common/entity/BaseEntity.java
@@ -1,0 +1,21 @@
+package com.common.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.PrePersist;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+public class BaseEntity {
+
+    @Column(nullable = false)
+    protected LocalDateTime createdAt;
+
+    @PrePersist
+    public void setCreatedAtNow() {
+        this.createdAt = LocalDateTime.now();
+    }
+}
