+++ src/main/java/com/member/entity/MemberEntity.java
@@ -1,18 +1,17 @@
 package com.member.entity;
 
+import com.common.entity.SoftDeletedEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Table(name = "member")
-public class MemberEntity {
+public class MemberEntity extends SoftDeletedEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,12 +27,6 @@ public class MemberEntity {
     @Column(nullable = false, unique = true)
     private String nickName;
 
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private boolean isDeleted = false;
-
     @Builder
     public MemberEntity(String email, String password, String nickName) {
         this.email = email;
@@ -46,14 +39,5 @@ public class MemberEntity {
         this.email = email;
         this.password = password;
         this.nickName = nickName;
-    }
-
-    public void softDelete() {
-        this.isDeleted = true;
-    }
-
-    @PrePersist
-    public void setCreatedAtNow() {
-        this.createdAt = LocalDateTime.now();
     }
 }
